Fix precision loss in numerical statistical indicators
This commit addresses issue #528 by removing premature casting to DoubleType
in statistical analyzers, which was causing precision loss for high-precision
numeric types like Decimal and large Long values.

Changes:
- Mean.scala: Remove .cast(DoubleType) from sum aggregation
- Sum.scala: Remove .cast(DoubleType) from sum aggregation
- RatioOfSums.scala: Remove .cast(DoubleType) from both sum aggregations
- Maximum.scala: Remove .cast(DoubleType) from max aggregation
- Minimum.scala: Remove .cast(DoubleType) from min aggregation

Impact:
By deferring the type conversion to Double until extraction time (via
result.getDouble()), Spark now maintains native precision during the
aggregation phase. This significantly improves precision for:
- Decimal/BigDecimal columns (e.g., financial data)
- Large numeric values that exceed Double's 53-bit mantissa precision
- Sum and mean calculations on high-precision data

The conversion to Double still occurs when extracting values via
result.getDouble(offset) in fromAggregationResult(), maintaining
API compatibility while improving precision during computation.

Fixes #528

+++ src/main/scala/com/amazon/deequ/analyzers/Maximum.scala
@@ -24,7 +24,6 @@ import com.google.common.annotations.VisibleForTesting
 import org.apache.spark.sql.functions.col
 import org.apache.spark.sql.functions.element_at
 import org.apache.spark.sql.functions.max
-import org.apache.spark.sql.types.DoubleType
 import org.apache.spark.sql.types.StructType
 import org.apache.spark.sql.Column
 import org.apache.spark.sql.Row
@@ -49,7 +48,7 @@ case class Maximum(column: String, where: Option[String] = None, analyzerOptions
     // The criterion returns a column where each row contains an array of 2 elements.
     // The first element of the array is a string that indicates if the row is "in scope" or "filtered" out.
     // The second element is the value used for calculating the metric. We use "element_at" to extract it.
-    max(element_at(criterion, 2).cast(DoubleType)) :: Nil
+    max(element_at(criterion, 2)) :: Nil
   }
 
   override def fromAggregationResult(result: Row, offset: Int): Option[MaxState] = {
