+++ src/main/scala/com/amazon/deequ/analyzers/Mean.scala
@@ -19,7 +19,7 @@ package com.amazon.deequ.analyzers
 import com.amazon.deequ.analyzers.Preconditions.{hasColumn, isNumeric}
 import org.apache.spark.sql.{Column, Row}
 import org.apache.spark.sql.functions.{count, sum}
-import org.apache.spark.sql.types.{DoubleType, StructType, LongType}
+import org.apache.spark.sql.types.{StructType, LongType}
 import Analyzers._
 
 case class MeanState(sum: Double, count: Long) extends DoubleValuedState[MeanState] {
@@ -38,7 +38,7 @@ case class Mean(column: String, where: Option[String] = None)
   with FilterableAnalyzer {
 
   override def aggregationFunctions(): Seq[Column] = {
-    sum(conditionalSelection(column, where)).cast(DoubleType) ::
+    sum(conditionalSelection(column, where)) ::
       count(conditionalSelection(column, where)).cast(LongType) :: Nil
   }
